requirements.txt
altgraph==0.15
certifi==2018.4.16
chardet==3.0.4
dnspython==1.15.0
docopt==0.6.2
et-xmlfile==1.0.1
future==0.16.0
idna==2.7
jdcal==1.4
macholib==1.9
modulegraph==0.16
odfpy==1.3.6
openpyxl==2.5.4
pefile==2017.11.5
pycrypto==2.6.1
PyInstaller==3.3.1
PyQt5==5.10.1
PyYAML==3.12
qqwry-py3==1.0.8
records==0.5.2
requests==2.19.1
sip==4.19.8
SQLAlchemy==1.2.9
tablib==0.12.1
tinydb==3.9.0.post1
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
